src/main/java/br/com/damsete/domain/page/PagedRequest.java
package br.com.damsete.domain.page;

import org.apache.commons.lang3.builder.ToStringBuilder;
import org.apache.commons.lang3.builder.ToStringStyle;

public class PagedRequest {

    private final int page;
    private final int size;
    private final String sortBy;
    private final String sortDir;
    private final String query;

    public PagedRequest(int page, int size, String sortBy, String sortDir, String query) {
        this.page = page;
        this.size = size;
        this.sortBy = sortBy;
        this.sortDir = sortDir;
        this.query = query;
    }

    public int getPage() {
        return page;
    }

    public int getSize() {
        return size;
    }

    public String getSortBy() {
        return sortBy;
    }

    public String getSortDir() {
        return sortDir;
    }

    public String getQuery() {
        return query;
    }

    @Override
    public String toString() {
        return ToStringBuilder.reflectionToString(this, ToStringStyle.SHORT_PREFIX_STYLE);
    }
}
